+++ example/lib/main.dart
@@ -4,6 +4,7 @@ import 'package:better_player_example/video_list/video_list_page.dart';
 import 'package:flutter/material.dart';
 
 import 'general_page/general_page.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
 
 void main() => runApp(MyApp());
 
@@ -13,6 +14,14 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Better player demo',
+      localizationsDelegates: [
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+      ],
+      supportedLocales: [
+        const Locale('en', 'US'),
+        const Locale('pl', 'PL'),
+      ],
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),