+++ lib/src/controls/better_player_material_controls.dart
@@ -57,7 +57,11 @@ class _BetterPlayerMaterialControlsState
         _cancelAndRestartTimer();
       },
       child: GestureDetector(
-        onTap: () => _cancelAndRestartTimer(),
+        onTap: _cancelAndRestartTimer,
+        onDoubleTap: () {
+          _cancelAndRestartTimer();
+          _onPlayPause();
+        },
         child: AbsorbPointer(
           absorbing: _hideStuff,
           child: Column(
@@ -116,7 +120,7 @@ class _BetterPlayerMaterialControlsState
               size: 42,
             ),
             Text(
-              _controlsConfiguration.defaultErrorText,
+              _betterPlayerController.translations.generalDefaultError,
               style: TextStyle(color: _controlsConfiguration.textColor),
             ),
           ],
@@ -125,21 +129,23 @@ class _BetterPlayerMaterialControlsState
     }
   }
 
-  AnimatedOpacity _buildTopBar() {
-    return AnimatedOpacity(
-      opacity: _hideStuff ? 0.0 : 1.0,
-      duration: _controlsConfiguration.controlsHideTime,
-      onEnd: _onPlayerHide,
-      child: Container(
-        height: _controlsConfiguration.controlBarHeight,
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.end,
-          children: [
-            _buildMoreButton(),
-          ],
-        ),
-      ),
-    );
+  Widget _buildTopBar() {
+    return _controlsConfiguration.enableOverflowMenu
+        ? AnimatedOpacity(
+            opacity: _hideStuff ? 0.0 : 1.0,
+            duration: _controlsConfiguration.controlsHideTime,
+            onEnd: _onPlayerHide,
+            child: Container(
+              height: _controlsConfiguration.controlBarHeight,
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.end,
+                children: [
+                  _buildMoreButton(),
+                ],
+              ),
+            ),
+          )
+        : const SizedBox();
   }
 
   Widget _buildMoreButton() {
@@ -147,8 +153,8 @@ class _BetterPlayerMaterialControlsState
       child: Padding(
         padding: const EdgeInsets.all(8),
         child: Icon(
-          Icons.more_vert,
-          color: Colors.white,
+          _controlsConfiguration.overflowMenuIcon,
+          color: _controlsConfiguration.iconsColor,
         ),
       ),
       onTap: () {
@@ -195,7 +201,7 @@ class _BetterPlayerMaterialControlsState
   Widget _buildLiveWidget() {
     return Expanded(
       child: Text(
-        _controlsConfiguration.liveText,
+        _betterPlayerController.translations.controlsLive,
         style: TextStyle(
             color: _controlsConfiguration.liveTextColor,
             fontWeight: FontWeight.bold),
@@ -315,7 +321,7 @@ class _BetterPlayerMaterialControlsState
                 child: Padding(
                   padding: const EdgeInsets.all(12),
                   child: Text(
-                    "Next video in ${snapshot.data} ...",
+                    "${_betterPlayerController.translations.controlsNextVideoIn} ${snapshot.data} ...",
                     style: TextStyle(color: Colors.white),
                   ),
                 ),