+++ lib/src/core/better_player_with_controls.dart
@@ -66,13 +66,23 @@ class _BetterPlayerWithControlsState extends State<BetterPlayerWithControls> {
     final BetterPlayerController betterPlayerController =
         BetterPlayerController.of(context);
 
+    var aspectRatio;
+    if (betterPlayerController.isFullScreen) {
+      aspectRatio = betterPlayerController
+              .betterPlayerConfiguration.fullScreenAspectRatio ??
+          BetterPlayerUtils.calculateAspectRatio(context);
+    } else {
+      aspectRatio =
+          betterPlayerController.betterPlayerConfiguration.aspectRatio ??
+              BetterPlayerUtils.calculateAspectRatio(context);
+    }
+
     return Center(
       child: Container(
         width: double.infinity,
         color: Colors.black,
         child: AspectRatio(
-          aspectRatio: betterPlayerController.aspectRatio ??
-              BetterPlayerUtils.calculateAspectRatio(context),
+          aspectRatio: aspectRatio,
           child: _buildPlayerWithControls(betterPlayerController, context),
         ),
       ),